refactor: внес корректировки на основе ревью
Вынес данные в попку consts/Questions, изменил компоненты верстки, удалил компоненты "Заглушки",
перенес папку RadioGroup в папку ui
