+++ src/widgets/Matches/Survey/consts/Questions/AestheticsQuestions.ts
@@ -1,0 +1,7 @@
+export const AestheticsQuestions = <const>[
+  'Ходил ли ты в театр за последние полгода?',
+  'Тебе интересна классическая музыка?',
+  'Посещал ли ты художественные выставки за последние полгода?',
+  'Тебе нравится посещать музеи?',
+  'Ты получаешь удовольствие от изучения архитектуры?'
+]
