+++ src/widgets/Matches/Survey/consts/Questions/HobbyQuestions.ts
@@ -1,0 +1,12 @@
+export const HobbyQuestions = <const>[
+  'Любишь ли ты смотреть кино?',
+  'Занимался ли ты рисованием?',
+  'Ты считаешь себя физически активным человеком?',
+  'Хорошо ли ты относишься к киберспорту?',
+  'Ты когда-нибудь занимался танцами?',
+  'Считаешь ли ты себя книголюбом?',
+  'Ты сочинял песни/музыку?',
+  'Ты получаешь удовольствие от приготовления еды?',
+  'Увлекался ли ты изучением иностранных языков?',
+  'Собирал ли ты коллекции чего-либо?'
+]
