+++ src/widgets/Matches/Survey/types/types.ts
@@ -1,0 +1,5 @@
+import {AestheticsQuestions} from '../consts/Questions/AestheticsQuestions'
+import {HobbyQuestions} from '../consts/Questions/HobbyQuestions'
+
+export type AestheticsQuestion = (typeof AestheticsQuestions)[number]
+export type HobbyQuestion = (typeof HobbyQuestions)[number]
