--- src/widgets/Matches/Survey/ui/Stages/Expectations/ChronoType/ChronoType.tsx
@@ -1,6 +1,0 @@
-import {ComponentType} from 'react'
-
-const ChronoType: ComponentType = () => {
-  return <div></div>
-}
-export default ChronoType
