--- src/widgets/Matches/Survey/ui/Stages/Expectations/Values/Values.tsx
@@ -1,6 +1,0 @@
-import {ComponentType} from 'react'
-
-const Values: ComponentType = () => {
-  return <div></div>
-}
-export default Values
