+++ src/widgets/Matches/Survey/ui/Stages/Interests/Aesthetics/Aesthetics.tsx
@@ -1,27 +1,22 @@
-import {Flex, Heading} from '@chakra-ui/react'
+import {Heading, VStack} from '@chakra-ui/react'
 import {ComponentType} from 'react'
 
-import RadioGroupQuestion from '@src/widgets/Matches/Survey/const/RadioGroup/RadioGroup'
+import {AestheticsQuestions} from '@src/widgets/Matches/Survey/consts/Questions/AestheticsQuestions'
+import RadioGroupQuestion from '@widgets/Matches/Survey/ui/RadioGroup/RadioGroup'
 
 const Aesthetics: ComponentType = () => {
-  const questions = [
-    'Ходил ли ты в театр за последние полгода?',
-    'Тебе интересна классическая музыка?',
-    'Посещал ли ты художественные выставки за последние полгода?',
-    'Тебе нравится посещать музеи?',
-    'Ты получаешь удовольствие от изучения архитектуры?'
-  ]
+  const questions = AestheticsQuestions
   return (
     <>
-      <Heading as='h3' textAlign='center' marginBottom='4.5'>
+      <Heading as='h3' textAlign='center' mb='4.5'>
         Эстетика
       </Heading>
 
-      <Flex as='div' wrap='wrap'>
-        {questions.map(question => {
-          return <RadioGroupQuestion question={question} />
+      <VStack as='div' wrap='wrap'>
+        {questions.map((question, key) => {
+          return <RadioGroupQuestion key={key} question={question} />
         })}
-      </Flex>
+      </VStack>
     </>
   )
 }
