+++ src/widgets/Matches/Survey/ui/Stages/Interests/Interests.tsx
@@ -1,4 +1,3 @@
-import {Heading} from '@chakra-ui/react'
 import {ComponentType} from 'react'
 
 import Hobby from './Hobby/Hobby'
@@ -6,9 +5,6 @@ import Hobby from './Hobby/Hobby'
 const Interests: ComponentType = () => {
   return (
     <>
-      <Heading as='h2' marginBottom='1.5rem' fontSize='22px'>
-        Интересы:
-      </Heading>
       <Hobby />
     </>
   )
