--- src/widgets/Matches/Survey/ui/Stages/PhysicalData/AgeInfo/AgeInfo.tsx
@@ -1,6 +1,0 @@
-import {ComponentType} from 'react'
-
-const AgeInfo: ComponentType = () => {
-  return <div></div>
-}
-export default AgeInfo
