--- src/widgets/Matches/Survey/ui/Stages/PhysicalData/PhysicalData.tsx
@@ -1,13 +1,0 @@
-import {Heading} from '@chakra-ui/react'
-import {ComponentType} from 'react'
-
-const PhysicalData: ComponentType = () => {
-  return (
-    <>
-      <Heading as='h2' marginBottom='1.5rem' fontSize='22px'>
-        Физические данные:
-      </Heading>
-    </>
-  )
-}
-export default PhysicalData
