--- src/widgets/Matches/Survey/ui/Stages/PhysicalData/SoftSkills/SoftSkills.tsx
@@ -1,6 +1,0 @@
-import {ComponentType} from 'react'
-
-const SoftSkills: ComponentType = () => {
-  return <div></div>
-}
-export default SoftSkills
