+++ src/widgets/Matches/Survey/ui/Survey/Survey.tsx
@@ -1,4 +1,4 @@
-import {FormControl, FormLabel, Button, Flex} from '@chakra-ui/react'
+import {FormControl, FormLabel, Button, Flex, Box} from '@chakra-ui/react'
 import {ComponentType} from 'react'
 
 import Interests from '../Stages/Interests/Interests'
@@ -6,14 +6,16 @@ import Interests from '../Stages/Interests/Interests'
 const Survey: ComponentType = () => {
   return (
     <FormControl as='fieldset'>
-      <Flex as='div' direction='column'>
+      <Box as='div'>
         <FormLabel as='form' w='100%' marginBottom='1.5rem'>
           <Interests />
         </FormLabel>
-        <Button alignSelf='flex-end' colorScheme='blue'>
-          Далее
-        </Button>
-      </Flex>
+        <Flex direction='column'>
+          <Button alignSelf='flex-end' colorScheme='blue'>
+            Далее
+          </Button>
+        </Flex>
+      </Box>
     </FormControl>
   )
 }
